requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
websockets==12.0
python-dotenv==1.0.1
langchain==0.2.5
langchain-anthropic==0.1.15
langchain-community==0.2.5
langgraph==0.1.19
langsmith==0.1.77
deepgram-sdk==3.2.7
elevenlabs==1.1.3
twilio==9.1.1
chromadb==0.5.0
sentence-transformers==3.0.1
pypdf==4.2.0
httpx==0.27.0
pydantic==2.7.3
pydantic-settings==2.3.0
streamlit==1.35.0
pandas==2.2.2
python-multipart==0.0.9
aiofiles==23.2.1
